+++ NWNXLib/API/API/CResRef.hpp
@@ -45,8 +45,7 @@ struct CResRef
     uint8_t GetLength() const;
 
     // Custom utility impl, missing from the API
-    // TODO: Commented out for now because this breaks stuff for some reason
-    // CResRef(const CResRef& other) { *this = other; }
+    CResRef(const CResRef& other) = default;
 
 #ifdef NWN_CLASS_EXTENSION_CResRef
     NWN_CLASS_EXTENSION_CResRef