+++ NWNXLib/nwnx.hpp
@@ -174,6 +174,18 @@ namespace String
     bool EndsWith(const std::string& str, const std::string& suffix);
 }
 
+namespace VectorMath
+{
+    float MagnitudeSquared(const Vector& v);
+    float Magnitude(const Vector& v);
+    float Dot(const Vector& a, const Vector& b);
+    Vector Add(const Vector& a, const Vector& b);
+    Vector Subtract(const Vector& a, const Vector& b);
+    Vector Multiply(const Vector& v, float s);
+    Vector Normalize(const Vector &v);
+    Vector Lineproject(const Vector &a, const Vector &b, const Vector &c);
+}
+
 namespace Utils
 {
     std::string ObjectIDToString(const ObjectID id);